Studio: show Documents & RAG in Data settings
The embedding model decides how uploaded documents are indexed, so it
belongs beside the files it applies to. Extract the section out of
General and render it in both, so the two cannot drift.
